invocacion a procedimientos/funciones

+++ lib/symbolTable/Symbol.java
@@ -24,7 +24,7 @@ abstract public class Symbol implements Cloneable {
     //será NONE para no parámetros
     public ParameterClass parClass; 
     //dirección en memoria. Para func/proc, dirección de la primera instrucción
-    public long dir; 
+    public int dir;
     //mi tipo
     public Types type;
     // es constante (En adac, siempre será FALSE)  
@@ -41,7 +41,7 @@ abstract public class Symbol implements Cloneable {
     	this (_name, -1, _type, _parClass, false); 
     }
    
-    public Symbol (String _name, long _dir, Types _type, ParameterClass _parClass, boolean _constant) {
+    public Symbol (String _name, int _dir, Types _type, ParameterClass _parClass, boolean _constant) {
     	name = _name;
     	dir = _dir; 
     	type = _type;
